updated elk script structure
