+++ monitoring/elk/lambdas/cloudwatch_ec2_metrics/cloudwatch_ec2_metrics.js
