+++ monitoring/elk/lambdas/elk_curator/elk_curator.py
