+++ monitoring/elk/template_mappings/cw_template.json
@@ -1,0 +1,21 @@
+{
+    "template": "cw-*",
+    "mappings": {
+        "_default_": {
+            "properties": {
+                "instance": {
+                    "index": "not_analyzed",
+                    "type": "string"
+                },
+                "instanceName": {
+                    "index": "not_analyzed",
+                    "type": "string"
+                },
+                "account": {
+                    "index": "not_analyzed",
+                    "type": "string"
+                }
+            }
+        }
+    }
+}
